Add original_file.id to FileSetIndexer
The `original_file.id` property is the id
you use to get the original file from Fedora.

Right now you need to get the id from the `FileSet`
object, which is not ideal.

For example, the `iiif_manifest` gem queries the
`FileSet` object for this directly, which causes
performance problems when generating a manifest for a
work that has child works.

If you had a `Book` that had child `Page` objects you
need the `original_file.id` for each `Page`'s `FileSet`.

If the `Book` has a thousand pages `::FileSet.original_file.id`
is going to be called 1000 times. Having this property
indexed in solr would allow for a solr query that gets that
information at once with a list of `Work` ids.